Move ColumnInfo outside of table to avoid circular import [skip ci]
