+++ astropy/coordinates/sky_coordinate.py
@@ -13,6 +13,7 @@ from ..units import Unit, IrreducibleUnit
 from .. import units as u
 from ..wcs.utils import skycoord_to_pixel, pixel_to_skycoord
 from ..utils.exceptions import AstropyDeprecationWarning
+from ..utils.column_info import ColumnInfo
 
 from .distances import Distance
 from .baseframe import BaseCoordinateFrame, frame_transform_graph, GenericFrame, _get_repr_cls
@@ -177,7 +178,6 @@ class SkyCoord(object):
         Mixin column information (attributes)
         """
         if not hasattr(self, '_info'):
-            from ..table import ColumnInfo
             self._info = ColumnInfo()
         return self._info
