+++ astropy/table/column.py
@@ -14,6 +14,7 @@ from ..units import Unit, Quantity
 from ..utils.compat import NUMPY_LT_1_8
 from ..utils.console import color_print
 from ..utils.metadata import MetaData
+from ..utils.column_info import COLUMN_ATTRS, ColumnInfo
 from . import groups
 from . import pprint
 from .np_utils import fix_column_name
@@ -42,60 +43,6 @@ _comparison_functions = set(
     [np.greater, np.greater_equal, np.less, np.less_equal,
      np.not_equal, np.equal,
      np.isfinite, np.isinf, np.isnan, np.sign, np.signbit])
-
-
-COLUMN_ATTRS = set(['name', 'unit', 'dtype', 'format', 'description', 'meta', 'parent_table'])
-
-class ColumnInfo(object):
-    cols_from_parent = set()
-
-    def __init__(self, parent_col=None):
-        self._attrs = dict((attr, None) for attr in COLUMN_ATTRS)
-        if parent_col is not None:
-            self._parent_col = weakref.ref(parent_col)
-
-    def __getattr__(self, attr):
-        if attr in self.cols_from_parent:
-            return getattr(self._parent_col(), attr)
-
-        try:
-            value = self._attrs[attr]
-        except KeyError:
-            super(ColumnInfo, self).__getattr__(attr)  # Generate AttributeError
-
-        # Weak ref for parent table
-        if attr == 'parent_table' and callable(value):
-            value = value()
-
-        # Mixins have a default dtype of Object if nothing else was set
-        if attr == 'dtype' and value is None:
-            value = np.dtype('O')
-
-        return value
-
-    def __setattr__(self, attr, value):
-        if attr in self.cols_from_parent:
-            setattr(self._parent_col(), attr, value)
-            return
-
-        if attr.startswith('_'):
-            super(ColumnInfo, self).__setattr__(attr, value)
-            return
-
-        if attr not in COLUMN_ATTRS:
-            raise AttributeError("attribute must be one of {0}".format(COLUMN_ATTRS))
-
-        if attr == 'parent_table':
-            value = None if value is None else weakref.ref(value)
-
-        self._attrs[attr] = value
-
-    def copy(self):
-        out = self.__class__()
-        for attr in COLUMN_ATTRS - self.cols_from_parent:
-            setattr(out, attr, deepcopy(getattr(self, attr)))
-
-        return out
 
 
 def col_setattr(col, attr, value):
